backend/api/requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
# Infrastructure communication (for MinIO/S3 and Kafka Producer)
boto3
confluent-kafka
sqlalchemy
psycopg2-binary
# Authentication (for passlib and python-jose)
passlib[bcrypt]
python-jose
# Infrastructure Clients
minio
confluent-kafka
PyJWT